Feat add manager factory and integrate with Tag
