+++ src/entities/core/index.ts
@@ -4,3 +4,4 @@ export * from "./utils";
 export * from "./types";
 export * from "./auth";
 export * from "./managerContract";
+export * from "./createManager";
